test(services) Add two tests for hadith tag services
Issue #245
