+++ HadithHouseWebsite/hadiths/exceptions.py
@@ -1,0 +1,9 @@
+class FacebookError(Exception):
+  """
+  Raised when there is a problem communicating with Facebook.
+  """
+  pass
+
+
+class MethodNotConfigured(Exception):
+  pass
